Populate cards from IndexedDB when first loaded
Adding the empty array as the list of dependencies to useEffect means this is only triggered once when the app is first loaded.
Settled for using setCards as it's simpler than looping and adding one at a time.
I did try looping through the cards and adding them using addNewCard but that created a warning about it being a dependency which needed to be in the list passed to useEffect but adding it to the list retriggered useEffect over and over again.