Add delete button to each card
This took a few goes to get right but in the end I created the actual code to remove the card, and it's record in IndexedDb if being used, into the App component.
This just felt like the right place in the end and because of this I thought I may as well move the code to add a card there too.